Support Get.isCheckExistenceOnly(). HBase 1.0 had a setting for only checking existence, which is intended to have the same functionality as Table.exists(Get).  `GetAdapter` should convert Gets with that flag to appropriate Bigtable Filter objects.  In addition, HBase 1 and 2 exist checks should just create new Get objects with `setCheckExistenceOnly(true)`. This should fix issue #1707.
